+++ CHANGELOG.md
@@ -5,4 +5,13 @@ A list of all the changes made to this repo
 Version 0.1.0
 ------------
 
+1. Added several extension attribute scripts
+    1. ARDStatus - Checks for Apple Remote Desktop status
+    2. FirmwarePasswordCheck - Checks if the firmware password was set
+    3. FusionSerialNumbers - Inventories the serial numbers tied to an installed VMware Fusion application
+    4. HomebrewInstalled - Checks if Homebrew is installed
+
+Version 0.1.0
+------------
+
 1. Initial Release of Homebrew script